requirements.txt
gdown==3.12.2
importlib-metadata==1.7.0
numpy==1.14.5
opencv-python==4.4.0.46
pre-commit==2.9.2
pytest==6.1.2
python-dotenv
torch==1.6.0
torchvision==0.2.1